refactor(theatron-core): extract shared API layer from TUI into core
Move HTTP client, SSE connection, streaming, error types, domain IDs,
and event definitions from theatron-tui into theatron-core so the
desktop app can reuse the API layer without depending on the TUI.

- api/client: full ApiClient with REST methods for sessions, agents, auth
- api/sse: SseConnection with auto-reconnect and exponential backoff
- api/streaming: per-turn SSE streaming with stream_message()
- api/error: snafu-based ApiError (Http, Server, Auth, InvalidToken)
- api/types: all request/response DTOs and SseEvent enum
- id: NousId, SessionId, TurnId, ToolId, PlanId newtypes
- events: StreamEvent enum with all turn/tool/plan variants

TUI retains thin re-export facades to preserve existing import paths.
Desktop updated to import shared types from core, removing its own
duplicated api/types module.

Closes P601
