+++ crates/theatron/core/Cargo.toml
@@ -4,13 +4,22 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
-description = "Shared presentation types and traits for Aletheia UIs"
+description = "Shared API client, types, SSE, and streaming infrastructure for Aletheia UIs"
 publish = false
 
 [dependencies]
+aletheia-koina = { path = "../../koina" }
 bytes = { workspace = true }
+compact_str = { workspace = true }
 futures-core = "0.3"
+futures-util = "0.3"
+reqwest = { workspace = true, features = ["stream", "cookies"] }
+rustls = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
+snafu = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
 
 [dev-dependencies]
 pulldown-cmark = "0.13"
