+++ crates/theatron/core/clippy.toml
@@ -1,5 +1,4 @@
-# theatron/core: shared type definitions for the theatron UI family.
-# Pure data types only — no network, database, or runtime dependencies.
+# theatron/core: shared API client, types, SSE, and streaming for Aletheia UIs.
 
 disallowed-methods = [
     { path = "std::process::exit", reason = "do not call process::exit from library code" },
@@ -11,8 +10,6 @@ disallowed-methods = [
 ]
 
 disallowed-types = [
-    # theatron/core is a type library; HTTP client belongs in the UI crate itself
-    { path = "reqwest::Client", reason = "theatron-core must not perform HTTP calls; use theatron-tui or theatron-desktop" },
-    # UI type library must not open DB connections
+    # UI library must not open DB connections
     { path = "rusqlite::Connection", reason = "theatron-core must not access SQLite; persistence belongs in mneme" },
 ]
