+++ crates/theatron/core/src/api/error.rs
@@ -1,11 +1,13 @@
+//! API-layer error type for the HTTP client.
+//!
+//! Covers three failure modes:
+//! - `Http`: a transport or connection error from `reqwest`
+//! - `Server`: a non-2xx HTTP response with a human-readable message from the server body
+//! - `Auth`: a 401/403 from the gateway
+
 use snafu::prelude::*;
 
-/// API-layer error type for the TUI HTTP client.
-///
-/// Covers three failure modes:
-/// - `Http`: a transport or connection error from `reqwest`
-/// - `Server`: a non-2xx HTTP response with a human-readable message from the server body
-/// - `Auth`: a 401/403 from the gateway
+/// Errors returned by [`super::ApiClient`] methods.
 #[derive(Debug, Snafu)]
 #[snafu(visibility(pub(crate)))]
 #[non_exhaustive]
@@ -13,14 +15,18 @@ pub enum ApiError {
     /// HTTP transport or connection error (no response received).
     #[snafu(display("{operation}: {source}"))]
     Http {
+        /// Which API call failed.
         operation: &'static str,
+        /// Underlying reqwest error.
         source: reqwest::Error,
     },
 
     /// Non-2xx HTTP response. Message is extracted from the server body when possible.
     #[snafu(display("{operation}: {message}"))]
     Server {
+        /// Which API call failed.
         operation: &'static str,
+        /// Human-readable error from the server.
         message: String,
     },
