+++ crates/theatron/desktop/src/api/mod.rs
@@ -2,4 +2,3 @@
 
 pub mod sse;
 pub mod streaming;
-pub mod types;
