+++ crates/theatron/tui/src/error.rs
@@ -11,7 +11,9 @@ use snafu::prelude::*;
 pub enum Error {
     /// API transport or authentication error from the HTTP client.
     #[snafu(context(false))]
-    Api { source: crate::api::error::ApiError },
+    Api {
+        source: theatron_core::api::ApiError,
+    },
 
     /// Token is required but was not supplied.
     #[snafu(display("{message}"))]
