+++ crates/theatron/tui/src/events.rs
@@ -1,5 +1,6 @@
-use crate::api::types::SseEvent;
-use crate::id::{NousId, PlanId, SessionId, ToolId, TurnId};
+use theatron_core::api::types::SseEvent;
+
+pub use theatron_core::events::StreamEvent;
 
 /// Raw events from the three multiplexed sources.
 /// Mapped to `Msg` by `App::map_event`.
@@ -14,66 +15,6 @@ pub enum Event {
     Stream(StreamEvent),
     /// 60fps UI tick
     Tick,
-}
-
-/// Parsed events from a POST /api/sessions/stream response
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum StreamEvent {
-    TurnStart {
-        session_id: SessionId,
-        nous_id: NousId,
-        turn_id: TurnId,
-    },
-    TextDelta(String),
-    ThinkingDelta(String),
-    ToolStart {
-        tool_name: String,
-        tool_id: ToolId,
-        input: Option<serde_json::Value>,
-    },
-    ToolResult {
-        tool_name: String,
-        tool_id: ToolId,
-        is_error: bool,
-        duration_ms: u64,
-        result: Option<String>,
-    },
-    ToolApprovalRequired {
-        turn_id: TurnId,
-        tool_name: String,
-        tool_id: ToolId,
-        input: serde_json::Value,
-        risk: String,
-        reason: String,
-    },
-    ToolApprovalResolved {
-        tool_id: ToolId,
-        decision: String,
-    },
-    PlanProposed {
-        plan: crate::api::types::Plan,
-    },
-    PlanStepStart {
-        plan_id: PlanId,
-        step_id: u32,
-    },
-    PlanStepComplete {
-        plan_id: PlanId,
-        step_id: u32,
-        status: String,
-    },
-    PlanComplete {
-        plan_id: PlanId,
-        status: String,
-    },
-    TurnComplete {
-        outcome: crate::api::types::TurnOutcome,
-    },
-    TurnAbort {
-        reason: String,
-    },
-    Error(String),
 }
 
 #[cfg(test)]
